Add new wpcom tests project
