+++ .teamcity/settings.kts
@@ -39,6 +39,7 @@ project {
 	vcsRoot(WpCalypso)
 	subProject(_self.projects.DesktopApp)
 	subProject(_self.projects.WPComPlugins)
+	subProject(_self.projects.WPComPlugins)
 	subProject(_self.projects.WebApp)
 	buildType(BuildBaseImages)
 	buildType(CheckCodeStyle)
